containerize bean item details
